+++ packages/core/src/actions/index.ts
@@ -6,6 +6,7 @@ export * from './get-ans-token-from-api'
 export * from './get-ans-tokens-from-api'
 export * from './get-modules-from-api'
 export * from './get-version-control-address-from-api'
+export * from './get-chain-rpc-from-api'
 export * from './get-sub-account-ids-by-module-id'
 
 export * from './account/public/get-account-base-addresses-from-api'