+++ packages/core/src/actions/public/types.ts
@@ -5,4 +5,5 @@ export enum CommonModuleNames {
   MODULE_FACTORY = 'module-factory',
   ANS_HOST = 'ans-host',
   IBC_CLIENT = 'ibc-client',
+  IBC_HOST = 'ibc-host',
 }