+++ packages/core/src/utils/cosmwasm/raw-query.ts
@@ -1,6 +1,11 @@
-import { type CosmWasmClient } from '@cosmjs/cosmwasm-stargate'
+import { CosmWasmClient } from '@cosmjs/cosmwasm-stargate'
 import { fromAscii, toAscii } from '@cosmjs/encoding'
 
+type RawQuery = {
+  cosmWasmClient: CosmWasmClient | undefined
+  address: string
+  key: string
+}
 /**
  * Perform a rawQuery on a contract.
  * @param readOnlyClient
@@ -8,16 +13,12 @@ import { fromAscii, toAscii } from '@cosmjs/encoding'
  * @param key - contract key like "contract_info"
  */
 export const rawQuery = async <TResponse = unknown>({
-  readOnlyClient,
+  cosmWasmClient,
   address,
   key,
-}: {
-  readOnlyClient: CosmWasmClient | undefined
-  address: string
-  key: string
-}): Promise<TResponse> => {
-  if (!readOnlyClient) return Promise.reject(new Error('No readonly client'))
-  return readOnlyClient
+}: RawQuery): Promise<TResponse> => {
+  if (!cosmWasmClient) return Promise.reject(new Error('No cosmWasmClient'))
+  return cosmWasmClient
     .queryContractRaw(address, toAscii(key))
     .then((response) => {
       if (!response) return response