+++ packages/react/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @abstract-money/react
 
+## 1.3.5
+
+### Patch Changes
+
+- Add remote simulation capabilities
+
+- Updated dependencies []:
+  - @abstract-money/core@1.3.2
+
 ## 1.3.4
 
 ### Patch Changes